Fix broken action buttons by removing them
Duplicate name check when adding a new student, click on the student to
view them inline in a modal dialog, click on the "Notes" button and it
breaks by redirecting the entire page instead of launching the notes
dialog.

Removing the buttons outright is correct here because the user shouldn't
be able to modify the user at this point, only view a quick summary.